src/requirements.txt
flask==1.1.2
python-dotenv==0.17.0
discord.py==1.7.1
discord==1.0.1